refactor!: rename GoogleOAuthClient to FireAuthClient
BREAKING CHANGE: GoogleOAuthClient has been renamed to FireAuthClient. Update any references accordingly.
