+++ fireauth2-server/src/web/extractors/fireauth.rs
@@ -1,20 +1,20 @@
 use std::ops::Deref;
 
-use fireauth2::GoogleOAuthClient;
+use fireauth2::FireAuthClient;
 
 use super::redirect_uri::RedirectUrl;
 
 #[derive(Clone)]
-pub struct FireAuth2(GoogleOAuthClient);
+pub struct FireAuth(FireAuthClient);
 
-impl Deref for FireAuth2 {
-    type Target = GoogleOAuthClient;
+impl Deref for FireAuth {
+    type Target = FireAuthClient;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
-impl actix_web::FromRequest for FireAuth2 {
+impl actix_web::FromRequest for FireAuth {
     type Error = actix_web::Error;
     type Future = futures::future::Ready<actix_web::Result<Self, Self::Error>>;
 
@@ -26,13 +26,13 @@ impl actix_web::FromRequest for FireAuth2 {
             .into_inner()
             .expect("RedirectUrl should be initialized on application startup");
 
-        match req.app_data::<actix_web::web::Data<GoogleOAuthClient>>() {
+        match req.app_data::<actix_web::web::Data<FireAuthClient>>() {
             Some(data) => {
                 let client = data
                     .as_ref()
                     .clone()
                     .with_redirect_uri(redirect_uri.inner_cloned());
-                futures::future::ok(FireAuth2(client))
+                futures::future::ok(FireAuth(client))
             }
             None => futures::future::err(
                 actix_web::error::ErrorInternalServerError(
